+++ packages/web-app/src/pages/newWithdraw.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState} from 'react';
 import {Address} from '@aragon/ui-components/dist/utils/addresses';
 import {useTranslation} from 'react-i18next';
 import {withTransaction} from '@elastic/apm-rum-react';
@@ -28,6 +28,7 @@ import {generatePath} from 'react-router-dom';
 import {useNetwork} from 'context/network';
 import {useDaoParam} from 'hooks/useDaoParam';
 import {Loading} from 'components/temporary';
+import {CreateProposalProvider} from 'context/createProposal';
 
 export type TokenFormData = {
   tokenName: string;
@@ -79,6 +80,7 @@ const NewWithdraw: React.FC = () => {
   const {t} = useTranslation();
   const {data: dao, loading} = useDaoParam();
   const {network} = useNetwork();
+  const [showTxModal, setShowTxModal] = useState(false);
 
   const formMethods = useForm<WithdrawFormData>({
     defaultValues,
@@ -144,53 +146,58 @@ const NewWithdraw: React.FC = () => {
   return (
     <>
       <FormProvider {...formMethods}>
-        <FullScreenStepper
-          wizardProcessName={t('newWithdraw.withdrawAssets')}
-          navLabel={t('allTransfer.newTransfer')}
-          returnPath={generatePath(Finance, {network, dao})}
+        <CreateProposalProvider
+          showTxModal={showTxModal}
+          setShowTxModal={setShowTxModal}
         >
-          <Step
-            wizardTitle={t('newWithdraw.configureWithdraw.title')}
-            wizardDescription={t('newWithdraw.configureWithdraw.subtitle')}
-            isNextButtonDisabled={
-              !configureWithdrawScreenIsValid(
-                dirtyFields.actions?.[0],
-                errors.actions?.[0],
-                tokenAddress
-              )
-            }
+          <FullScreenStepper
+            wizardProcessName={t('newWithdraw.withdrawAssets')}
+            navLabel={t('allTransfer.newTransfer')}
+            returnPath={generatePath(Finance, {network, dao})}
           >
-            <ConfigureWithdrawForm />
-          </Step>
-          <Step
-            wizardTitle={t('newWithdraw.setupVoting.title')}
-            wizardDescription={t('newWithdraw.setupVoting.description')}
-            isNextButtonDisabled={!setupVotingIsValid(errors, durationSwitch)}
-          >
-            <SetupVotingForm />
-          </Step>
-          <Step
-            wizardTitle={t('newWithdraw.defineProposal.heading')}
-            wizardDescription={t('newWithdraw.defineProposal.description')}
-            isNextButtonDisabled={!defineProposalIsValid(dirtyFields, errors)}
-          >
-            <DefineProposal />
-          </Step>
-          <Step
-            wizardTitle={t('newWithdraw.reviewProposal.heading')}
-            wizardDescription={t('newWithdraw.reviewProposal.description')}
-            nextButtonLabel={t('labels.submitWithdraw')}
-            isNextButtonDisabled
-            fullWidth
-          >
-            <ReviewProposal />
-          </Step>
-        </FullScreenStepper>
-        <TokenMenu
-          isWallet={false}
-          onTokenSelect={handleTokenSelect}
-          tokenBalances={balances}
-        />
+            <Step
+              wizardTitle={t('newWithdraw.configureWithdraw.title')}
+              wizardDescription={t('newWithdraw.configureWithdraw.subtitle')}
+              isNextButtonDisabled={
+                !configureWithdrawScreenIsValid(
+                  dirtyFields.actions?.[0],
+                  errors.actions?.[0],
+                  tokenAddress
+                )
+              }
+            >
+              <ConfigureWithdrawForm />
+            </Step>
+            <Step
+              wizardTitle={t('newWithdraw.setupVoting.title')}
+              wizardDescription={t('newWithdraw.setupVoting.description')}
+              isNextButtonDisabled={!setupVotingIsValid(errors, durationSwitch)}
+            >
+              <SetupVotingForm />
+            </Step>
+            <Step
+              wizardTitle={t('newWithdraw.defineProposal.heading')}
+              wizardDescription={t('newWithdraw.defineProposal.description')}
+              isNextButtonDisabled={!defineProposalIsValid(dirtyFields, errors)}
+            >
+              <DefineProposal />
+            </Step>
+            <Step
+              wizardTitle={t('newWithdraw.reviewProposal.heading')}
+              wizardDescription={t('newWithdraw.reviewProposal.description')}
+              nextButtonLabel={t('labels.submitWithdraw')}
+              onNextButtonClicked={() => setShowTxModal(true)}
+              fullWidth
+            >
+              <ReviewProposal />
+            </Step>
+          </FullScreenStepper>
+          <TokenMenu
+            isWallet={false}
+            onTokenSelect={handleTokenSelect}
+            tokenBalances={balances}
+          />
+        </CreateProposalProvider>
       </FormProvider>
     </>
   );
